requirements.txt
tensorflow
numpy
opencv-python
matplotlib
streamlit
pillow
nibabel
